refactor: update driver app packages and features
