+++ driver-app/app/proguard-rules.pro
@@ -1,3 +1,38 @@
--dontwarn kotlinx.serialization.**
+pro# Kotlinx Serialization
+-keepattributes *Annotation*, InnerClasses
+-dontnote kotlinx.serialization.AnnotationsKt
+-keepclassmembers class kotlinx.serialization.json.** {
+    *** Companion;
+}
+-keepclasseswithmembers class kotlinx.serialization.json.** {
+    kotlinx.serialization.KSerializer serializer(...);
+}
+-keep,includedescriptorclasses class com.yourco.driverAA.**$$serializer { *; }
+-keepclassmembers class com.yourco.driverAA.** {
+    *** Companion;
+}
+-keepclasseswithmembers class com.yourco.driverAA.** {
+    kotlinx.serialization.KSerializer serializer(...);
+}
+
+# Firebase
 -keep class com.google.firebase.** { *; }
 -keep class com.google.android.gms.** { *; }
+
+# Retrofit
+-keepattributes Signature, InnerClasses, EnclosingMethod
+-keepattributes RuntimeVisibleAnnotations, RuntimeVisibleParameterAnnotations
+-keepclassmembers,allowshrinking,allowobfuscation interface * {
+    @retrofit2.http.* <methods>;
+}
+-dontwarn org.codehaus.mojo.animal_sniffer.IgnoreJRERequirement
+-dontwarn javax.annotation.**
+-dontwarn kotlin.Unit
+-dontwarn retrofit2.KotlinExtensions*
+-if interface * { @retrofit2.http.* <methods>; }
+-keep,allowobfuscation interface <1>
+
+# OkHttp
+-dontwarn okhttp3.**
+-dontwarn okio.**
+
