--- driver-app/app/src/main/java/com/orderops/driver/MainActivity.kt
@@ -1,35 +1,0 @@
-package com.orderops.driver
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import dagger.hilt.android.AndroidEntryPoint
-import com.google.android.play.core.appupdate.AppUpdateManagerFactory
-import com.google.android.play.core.install.model.AppUpdateType
-import com.google.android.play.core.install.model.UpdateAvailability
-import com.orderops.driver.navigation.NavGraph
-
-@AndroidEntryPoint
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        checkForUpdates()
-        setContent {
-            MaterialTheme {
-                Surface { NavGraph() }
-            }
-        }
-    }
-
-    private fun checkForUpdates() {
-        val manager = AppUpdateManagerFactory.create(this)
-        manager.appUpdateInfo.addOnSuccessListener { info ->
-            if (info.updateAvailability() == UpdateAvailability.UPDATE_AVAILABLE &&
-                info.isUpdateTypeAllowed(AppUpdateType.FLEXIBLE)) {
-                manager.startUpdateFlowForResult(info, AppUpdateType.FLEXIBLE, this, 0)
-            }
-        }
-    }
-}
