--- driver-app/app/src/main/java/com/orderops/driver/data/db/AppDatabase.kt
@@ -1,9 +1,0 @@
-package com.orderops.driver.data.db
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-
-@Database(entities = [LocationPing::class], version = 1)
-abstract class AppDatabase : RoomDatabase() {
-    abstract fun locationDao(): LocationPingDao
-}
