--- driver-app/app/src/main/java/com/orderops/driver/domain/JobsRepository.kt
@@ -1,8 +1,0 @@
-package com.orderops.driver.domain
-
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOf
-
-class JobsRepository {
-    val jobs: Flow<List<String>> = flowOf(listOf("1", "2", "3"))
-}
