--- driver-app/app/src/main/java/com/orderops/driver/ui/jobdetail/JobDetailScreen.kt
@@ -1,9 +1,0 @@
-package com.orderops.driver.ui.jobdetail
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun JobDetailScreen(jobId: String) {
-    Text(text = "Job detail $jobId")
-}
