--- driver-app/app/src/main/java/com/orderops/driver/util/DeepLinks.kt
@@ -1,5 +1,0 @@
-package com.orderops.driver.util
-
-object DeepLinks {
-    const val JOB_BASE = "driver://job"
-}
