--- driver-app/app/src/main/java/com/orderops/driver/util/Result.kt
@@ -1,7 +1,0 @@
-package com.orderops.driver.util
-
-sealed class Result<out T> {
-    data class Success<T>(val data: T) : Result<T>()
-    data class Error(val throwable: Throwable) : Result<Nothing>()
-    object Loading : Result<Nothing>()
-}
