+++ driver-app/app/src/main/java/com/yourco/driverAA/data/db/AppDatabase.kt
@@ -1,0 +1,26 @@
+package com.yourco.driverAA.data.db
+
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
+import androidx.room.migration.Migration
+import androidx.sqlite.db.SupportSQLiteDatabase
+
+@Database(
+    entities = [LocationPing::class], 
+    version = 1,
+    exportSchema = false
+)
+abstract class AppDatabase : RoomDatabase() {
+    abstract fun locationDao(): LocationPingDao
+    
+    companion object {
+        val MIGRATION_1_2 = object : Migration(1, 2) {
+            override fun migrate(database: SupportSQLiteDatabase) {
+                // Example migration for future use
+                // database.execSQL("ALTER TABLE LocationPing ADD COLUMN newColumn TEXT")
+            }
+        }
+    }
+}
+
