+++ driver-app/app/src/main/java/com/yourco/driverAA/ui/jobdetail/JobDetailScreen.kt
@@ -1,0 +1,9 @@
+package com.yourco.driverAA.ui.jobdetail
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun JobDetailScreen(jobId: String) {
+    Text(text = "Job detail $jobId")
+}
