+++ driver-app/app/src/main/java/com/yourco/driverAA/util/DeepLinks.kt
@@ -1,0 +1,5 @@
+package com.yourco.driverAA.util
+
+object DeepLinks {
+    const val JOB_BASE = "driver://job"
+}
