+++ driver-app/app/src/main/java/com/yourco/driverAA/util/Result.kt
@@ -1,0 +1,7 @@
+package com.yourco.driverAA.util
+
+sealed class Result<out T> {
+    data class Success<T>(val data: T) : Result<T>()
+    data class Error(val throwable: Throwable) : Result<Nothing>()
+    object Loading : Result<Nothing>()
+}
