+++ driver-app/app/src/main/java/com/yourco/driverAA/work/WorkScheduling.kt
@@ -1,0 +1,24 @@
+package com.yourco.driverAA.work
+
+import android.content.Context
+import androidx.work.*
+import java.util.concurrent.TimeUnit
+
+object WorkScheduling {
+    fun scheduleUpload(context: Context) {
+        val request = PeriodicWorkRequestBuilder<UploadLocationWorker>(15, TimeUnit.MINUTES)
+            .setConstraints(
+                Constraints.Builder()
+                    .setRequiredNetworkType(NetworkType.CONNECTED)
+                    .build()
+            )
+            .addTag("upload-locations")
+            .build()
+        WorkManager.getInstance(context).enqueueUniquePeriodicWork(
+            "upload-locations",
+            ExistingPeriodicWorkPolicy.UPDATE,
+            request
+        )
+    }
+}
+
